Move game_files.js!
