+++ src/CharacterThread.js
@@ -3,7 +3,7 @@ const io = require("socket.io-client");
 const fs = require("fs").promises;
 const { JSDOM } = require("jsdom");
 const node_query = require("jquery");
-const game_files = require("../game_files");
+const game_files = require("./game_files");
 const fetch = (...args) =>
   import("node-fetch").then(({ default: fetch }) => fetch(...args));
 const monitoring_util = require("../monitoring_util");
